refactor(infra-board): move resolveSigningUrl out of the component file
SigningPanel.tsx exported both the component and a plain helper, which
breaks Fast Refresh and tripped react-refresh/only-export-components. The
rule's own remedy is a separate module, and that is what the helper wanted
anyway: it was exported solely so it could be unit-tested directly against
the actual iframe src rather than an intermediate variable.

The function body and its tests move verbatim -- no behaviour changes, and
the same 18 assertions run, now split across the two files. The one
paragraph of the doc comment that justified the export had gone stale in a
standalone module, so it now records both reasons the helper lives apart:
Fast Refresh, and direct testability of the src an iframe receives.
